requirements.txt
altair==5.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
bqplot==0.12.43
branca==0.8.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
colour==0.1.5
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
deep-translator==1.11.4
distlib==0.3.7
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
earthengine-api==1.1.2
eerepr==0.0.4
et-xmlfile==1.1.0
executing==2.1.0
Farama-Notifications==0.0.4
fastjsonschema==2.21.1
filelock==3.12.2
Flask==3.0.3
Flask-SocketIO==5.4.1
folium==0.17.0
fonttools==4.54.1
future==1.0.0
geemap==0.34.5
geocoder==1.38.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
gymnasium==1.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib-metadata==7.0.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.3
ipyleaflet==0.19.2
ipython==8.28.0
ipytree==0.2.2
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-leaflet==0.19.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mysql-connector-python==8.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.1.1
openpyxl==3.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==3.9.1
plotly==5.24.1
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.8.1b0
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.25.1
pyparsing==3.1.4
pyperclip==1.9.0
pyshp==2.3.1
python-box==7.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.2
pywin32==306
pyzmq==25.1.2
ratelim==0.1.6
referencing==0.32.1
reportlab==4.0.9
requests==2.32.3
rest-framework-simplejwt==0.0.2
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
scooby==0.10.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.30.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.9.0
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
validators==0.22.0
virtualenv==20.24.0
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wordcloud==1.9.4
wsproto==1.2.0
xyzservices==2024.9.0
zipp==3.17.0
